+++ components/atoms/subHeadingSpan/index.ts
