+++ components/atoms/subHeadingSpan/subHeadingSpan.styled.ts
