+++ components/organisms/setting/setting.tsx
@@ -3,8 +3,8 @@ import {useRouter} from "next/router";
 import React, {useCallback} from "react";
 import * as S from "./setting.styled";
 import {SliderSwitch} from "components/atoms/sliderSwitch";
-import {SubHeadingButton} from "components/subHeading/subHeadingButton";
-import {SubHeadingSpan} from "components/subHeading/subHeadingSpan";
+import {SubHeadingButton} from "components/atoms/subHeadingButton";
+import {SubHeadingSpan} from "components/atoms/subHeadingSpan";
 import {UserData} from "stores/data";
 
 type SettingProps = {
