--- components/socialLoginButton/index.ts
@@ -1,2 +1,0 @@
-export * from "./socialLoginButton"
-export * as S from "./socialLoginButton.styled"
