--- components/socialLoginButton/socialLoginButton.styled.ts
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-export const Button = styled.button`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 18vw;
-    height:4rem;
-    min-width: 300px;
-    max-height: 80px;
-    border-radius: 5px;
-    margin-bottom: 0.8rem;
-    cursor: pointer;
-
-    > span{
-        font-size: 1rem;
-    }
-`
