--- components/socialLoginButton/socialLoginButton.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-const SocialLoginButton = () => <h1>socialLoginButton</h1>;
-
-export default SocialLoginButton;
